publisher/requirements.txt
google-cloud-pubsub
requests
Flask
gunicorn
